fix: conv from float to int ph

+++ pi4/frontend/models.py
@@ -23,7 +23,7 @@ class MotorMessage(db.Model):
 
 class PhMessage(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    payload = db.Column(db.Float, nullable=False)
+    payload = db.Column(db.Integer, nullable=False)
     timestamp = db.Column(db.DateTime, default=db.func.current_timestamp())
 
 class AvailMessage(db.Model):
